+++ packages/block-library/src/video/edit.js
@@ -94,12 +94,22 @@ function VideoEdit( {
 			// in this case there was an error
 			// previous attributes should be removed
 			// because they may be temporary blob urls
-			setAttributes( { src: undefined, id: undefined } );
+			setAttributes( {
+				src: undefined,
+				id: undefined,
+				poster: undefined,
+			} );
 			return;
 		}
+
 		// sets the block's attribute and updates the edit component from the
 		// selected media
-		setAttributes( { src: media.url, id: media.id } );
+		setAttributes( {
+			src: media.url,
+			id: media.id,
+			poster:
+				media.image?.src !== media.icon ? media.image?.src : undefined,
+		} );
 	}
 
 	function onSelectURL( newSrc ) {
@@ -112,7 +122,7 @@ function VideoEdit( {
 				onReplace( embedBlock );
 				return;
 			}
-			setAttributes( { src: newSrc, id: undefined } );
+			setAttributes( { src: newSrc, id: undefined, poster: undefined } );
 		}
 	}
 
@@ -151,7 +161,7 @@ function VideoEdit( {
 	}
 
 	function onRemovePoster() {
-		setAttributes( { poster: '' } );
+		setAttributes( { poster: undefined } );
 
 		// Move focus back to the Media Upload button.
 		posterImageButton.current.focus();