+++ packages/components/src/toggle-group-control/types.ts
@@ -29,7 +29,7 @@ export type ToggleGroupControlProps = Omit<
 	 *
 	 * @default false
 	 */
-	hideLabelFromVision: boolean;
+	hideLabelFromVision?: boolean;
 	/**
 	 * Determines if segments should be rendered with equal widths.
 	 *
@@ -88,5 +88,6 @@ export type ToggleGroupControlButtonProps = {
 export type ToggleGroupControlBackdropProps = {
 	containerRef: MutableRefObject< HTMLElement | undefined >;
 	containerWidth?: number | null;
+	isAdaptiveWidth?: boolean;
 	state?: any;
 };