+++ packages/core-data/src/resolvers.js
@@ -27,8 +27,6 @@ import {
 	receiveEntityRecords,
 	receiveThemeSupports,
 	receiveEmbedPreview,
-	receiveUserPermission,
-	receiveAutosaves,
 } from './actions';
 import { getKindEntities, DEFAULT_ENTITY_KEY } from './entities';
 import { ifNotResolved, getNormalizedCommaSeparable } from './utils';
@@ -292,7 +290,7 @@ export function* getEmbedPreview( url ) {
  * @param {string}  resource REST resource to check, e.g. 'media' or 'posts'.
  * @param {?string} id       ID of the rest resource to check.
  */
-export function* canUser( action, resource, id ) {
+export const canUser = ( action, resource, id ) => async ( { dispatch } ) => {
 	const methods = {
 		create: 'POST',
 		read: 'GET',
@@ -309,7 +307,7 @@ export function* canUser( action, resource, id ) {
 
 	let response;
 	try {
-		response = yield apiFetch( {
+		response = await triggerFetch( {
 			path,
 			// Ideally this would always be an OPTIONS request, but unfortunately there's
 			// a bug in the REST API which causes the Allow header to not be sent on
@@ -337,8 +335,8 @@ export function* canUser( action, resource, id ) {
 
 	const key = compact( [ action, resource, id ] ).join( '/' );
 	const isAllowed = includes( allowHeader, method );
-	yield receiveUserPermission( key, isAllowed );
-}
+	dispatch.receiveUserPermission( key, isAllowed );
+};
 
 /**
  * Checks whether the current user can perform the given action on the given
@@ -348,16 +346,18 @@ export function* canUser( action, resource, id ) {
  * @param {string} name     Entity name.
  * @param {string} recordId Record's id.
  */
-export function* canUserEditEntityRecord( kind, name, recordId ) {
-	const entities = yield getKindEntities( kind );
+export const canUserEditEntityRecord = ( kind, name, recordId ) => async ( {
+	dispatch,
+} ) => {
+	const entities = await dispatch( getKindEntities( kind ) );
 	const entity = find( entities, { kind, name } );
 	if ( ! entity ) {
 		return;
 	}
 
 	const resource = entity.__unstable_rest_base;
-	yield canUser( 'update', resource, recordId );
-}
+	await dispatch( canUser( 'update', resource, recordId ) );
+};
 
 /**
  * Request autosave data from the REST API.
@@ -365,20 +365,19 @@ export function* canUserEditEntityRecord( kind, name, recordId ) {
  * @param {string} postType The type of the parent post.
  * @param {number} postId   The id of the parent post.
  */
-export function* getAutosaves( postType, postId ) {
-	const { rest_base: restBase } = yield controls.resolveSelect(
-		STORE_NAME,
-		'getPostType',
-		postType
-	);
-	const autosaves = yield apiFetch( {
+export const getAutosaves = ( postType, postId ) => async ( {
+	dispatch,
+	resolveSelect,
+} ) => {
+	const { rest_base: restBase } = await resolveSelect.getPostType( postType );
+	const autosaves = await triggerFetch( {
 		path: `/wp/v2/${ restBase }/${ postId }/autosaves?context=edit`,
 	} );
 
 	if ( autosaves && autosaves.length ) {
-		yield receiveAutosaves( postId, autosaves );
+		dispatch.receiveAutosaves( postId, autosaves );
 	}
-}
+};
 
 /**
  * Request autosave data from the REST API.
@@ -389,14 +388,11 @@ export function* getAutosaves( postType, postId ) {
  * @param {string} postType The type of the parent post.
  * @param {number} postId   The id of the parent post.
  */
-export function* getAutosave( postType, postId ) {
-	yield controls.resolveSelect(
-		STORE_NAME,
-		'getAutosaves',
-		postType,
-		postId
-	);
-}
+export const getAutosave = ( postType, postId ) => async ( {
+	resolveSelect,
+} ) => {
+	await resolveSelect.getAutosaves( postType, postId );
+};
 
 /**
  * Retrieve the frontend template used for a given link.